Add Sign-up & Log-in system to allow only the author delete his/her post

+++ app.js
@@ -3,10 +3,12 @@
 //Acquiring Dependencies-
 const express = require("express");
 const bodyParser = require("body-parser");
+const cookieParser = require("cookie-parser");
 const ejs = require("ejs");
 const mongoose = require('mongoose');
 const _ = require("lodash");
 const PORT = process.env.PORT || 3000;
+const auth = require('./middlewares/auth');
 
 //Default Texts-
 const homeStartingContent = "I'm Daily Journal, your best pal. What do I do? Well, I'm here to help you out. I'll be there to listen to your thoughts or share with you my pal's ideas and few amazing blogs.That's all? Not yet. I'm here to take you on a wonderful journey of unlimited thoughts and help you find your twin souls too!!! Sounds great? Here we go....Let's get started.";
@@ -18,7 +20,8 @@ const app = express();
 app.set('view engine', 'ejs');
 app.use(bodyParser.urlencoded({extended: true}));
 app.use(express.static("public"));
-
+app.use(express.json());
+app.use(cookieParser());
 //When in development mode then only require the dotenv module
 if(process.env.NODE_ENV !== 'production'){
   const dotenv = require('dotenv');
@@ -37,14 +40,22 @@ const blogSchema = {
   timestamps: {
     type: Date,
     default: Date.now
+  },
+  author: {
+    type: mongoose.Schema.Types.ObjectId,
+    ref: "User"
   }
 }
 
 //Making a MongoDB model for the schema-
 const Blog = new mongoose.model("Blog", blogSchema);
 
+// Router for user login and sign in
+app.use(require("./routes/user.router"));
+
 //Get request for home route-
-app.get(["/", "/page/:page", "/page/:perPage", "/page/:page/:perPage"], function(req, res){
+app.get(["/", "/page/:page", "/page/:perPage", "/page/:page/:perPage"], auth, function(req, res){
+  
   var perPage = parseInt(req.params.perPage) || 5;
   if(req.query.perPage>0)
     perPage =  parseInt(req.query.perPage);
@@ -67,7 +78,8 @@ app.get(["/", "/page/:page", "/page/:perPage", "/page/:page/:perPage"], function
           pages: Math.ceil(count/perPage),
           search: "",
           perPage: perPage,
-          order: order
+          order: order,
+          isAuthenticated: req.user? true: false
         });
       }
     })
@@ -76,28 +88,46 @@ app.get(["/", "/page/:page", "/page/:perPage", "/page/:page/:perPage"], function
 
 
 //Get request for about page-
-app.get("/about", function(req, res){
-  res.render("about", {aboutContent: aboutContent});
+app.get("/about", auth, function(req, res){
+  
+  res.render("about", {
+    aboutContent: aboutContent,
+    isAuthenticated: req.user? true : false
+  });
 });
 
 //Get request for contact page-
-app.get("/contact", function(req, res){
-  res.render("contact", {contactContent: contactContent});
+app.get("/contact", auth, function(req, res){
+  res.render("contact", {
+    contactContent: contactContent,
+    isAuthenticated: req.user? true : false
+  });
 });
 
 //Get request for compose blog page-
-app.get("/compose", function(req, res){
-  res.render("compose");
+app.get("/compose", auth, function(req, res){
+  const user = req.user;
+  if(!user){
+    return res.status(401).redirect("/log-in");
+  }
+  res.render("compose", {
+    isAuthenticated: true
+  });
 });
 
 //Post request to save the new blogs to the DB
-app.post("/compose", function(req, res){
+app.post("/compose", auth, function(req, res){
+  const user = req.user;
+  if(!user){
+    return res.status(401).redirect("/log-in");
+  }
   const postTitle = req.body.postTitle;
   const postContent = req.body.postBody;
   const blog = new Blog({
     blogTitle: postTitle,
     blogContent: postContent,
-    comments: []
+    comments: [],
+    author: user._id
   })
   console.log(blog);
   blog.save();
@@ -105,20 +135,28 @@ app.post("/compose", function(req, res){
 });
 
 //Get request for posts page-
-app.get(["/posts/:postName", "/page/posts/:postName", "/page/:page/posts/:postName", "/search/:query/posts/:postName", "/search/:query/:page/posts/:postName"], function(req, res){
+app.get(["/posts/:postName", "/page/posts/:postName", "/page/:page/posts/:postName", "/search/:query/posts/:postName", "/search/:query/:page/posts/:postName"], auth, function(req, res){
+  const user = req.user;
+  let isAuthor = false;
   const requestedTitle = _.lowerCase(req.params.postName);
   Blog.find({}, function(err, posts){
     if(!err){
       posts.forEach(function(post){
         const storedTitle = _.lowerCase(post.blogTitle);
         if(storedTitle === requestedTitle){
+          // Check if the user and author of this post are same
+          if(user && JSON.stringify(user._id) === JSON.stringify(post.author)){
+            isAuthor = true;
+          }
           //Sort the comments to show the recent one
           post.comments = post.comments.sort((a,b) =>  ((a.timestamps > b.timestamps) ? -1 : ((a.timestamps < b.timestamps) ? 1 : 0)));
           res.render("post", {
             title: post.blogTitle,
             content: post.blogContent,
             id:post._id,
-            comments: post.comments
+            comments: post.comments,
+            isAuthor,
+            isAuthenticated: user? true: false
           });
         }
       });
@@ -151,7 +189,7 @@ app.post("/posts/:postName/comment", async function(req, res) {
 });
 
 //Post request to search by title
-app.post(["/search"], function(req, res){
+app.post(["/search"], auth, function(req, res){
   const query = req.body.query || req.params.query;
   var perPage = 5;
   const currentPage = req.params.page || 1;
@@ -169,7 +207,8 @@ app.post(["/search"], function(req, res){
         pages: Math.ceil(count/perPage),
         search: query,
         perPage: perPage,
-        order: 'new one first'
+        order: 'new one first',
+        isAuthenticated: req.user? true : false
       });
     })
 
@@ -177,7 +216,7 @@ app.post(["/search"], function(req, res){
 })
 
 // GET request for search to support pagination
-app.get(["/search/:query/:page", "/search/:query", "/search/:query/:page/:perPage", "/search/:query"], function(req, res){
+app.get(["/search/:query/:page", "/search/:query", "/search/:query/:page/:perPage", "/search/:query"], auth, function(req, res){
   const query = req.params.query;
   var perPage = parseInt(req.params.perPage) || 5;
   if(req.query.perPage>0)
@@ -198,7 +237,8 @@ app.get(["/search/:query/:page", "/search/:query", "/search/:query/:page/:perPag
         pages: Math.ceil(count/perPage),
         search: query,
         perPage: perPage,
-        order: order
+        order: order,
+        isAuthenticated: req.user? true: false
       });
     })
 
@@ -206,10 +246,15 @@ app.get(["/search/:query/:page", "/search/:query", "/search/:query/:page/:perPag
 })
 
 //delete post route
-app.post('/posts/:postName', (req, res, next) => {
+app.post('/posts/:postName', auth, (req, res, next) => {
+  const user = req.user;
+  if(!user) {
+    return res.status(401).redirect("/log-in");
+  }
+
   const requestedTitle = req.params.postName;
   console.log(requestedTitle)
-  Blog.deleteOne({blogTitle: requestedTitle}).then(
+  Blog.deleteOne({blogTitle: requestedTitle, author: user._id}).then(
     () => {
       res.redirect("/");
     }
